Moved test plotting of filters into an example
