+++ tests/scattering1d/test_numpy_scattering1d.py
@@ -67,30 +67,12 @@ class TestScattering1DNumpy:
         # should we add this new result to the test data?
         #print('Sx1 shape: ' + str(Sx1.shape) + ' Sx2.shape: ' + str(Sx2.shape))
         assert Sx2.shape == (Sx1.shape[0], Sx1.shape[1], Sx1.shape[2]*2**(sigma_low_scale_factor))
-    
-    def Scattering1D_filter_factory_T_plot(self, N, J, T, phi_f, psi1_f, msg):
-          import matplotlib.pyplot as plt
-          plot_dir = 'plots'
-          if not os.path.isdir(plot_dir):
-              os.mkdir(plot_dir)    
-
-          plt.figure()
-          plt.plot(np.arange(N)/N, phi_f[0], 'r', label=msg)
-          for psi_f in psi1_f:
-              plt.plot(np.arange(N)/N, psi_f[0], 'b')
-          plt.xlim(0, 0.5)
-          plt.xlabel(r'$\omega$', fontsize=18)
-          plt.ylabel(r'$\hat\psi_j(\omega)$', fontsize=18)
-          
-          plt.legend()
-          plt.savefig(plot_dir + '/order1filters_T' + str(T) + 'J' + str(J) +  '.pdf', bbox_inches='tight', orientation='landscape')
-    
+        
     def test_Scattering1D_filter_factory_T(self, backend):
         """
         Constructs the scattering filters for the T parameter which controls the
         temporal extent of the low-pass sigma_log filter
         """
-        DO_PLOTS = False
         N = 2**13
         Q = 1       
         sigma_low_scale_factor = [0, 5]        
@@ -109,12 +91,3 @@ class TestScattering1DNumpy:
                 #print(msg)
                 assert(phi_f['sigma']==0.1/T)
 
-                if DO_PLOTS:
-                    self.Scattering1D_filter_factory_T_plot(N, J, T, phi_f, psi1_f, msg)
-                  
-if __name__ == '__main__':
-  test = TestScattering1DNumpy()
-#  test.test_Scattering1D_T(backend)
-  #test.test_Scattering1D(backend)
-  test.test_Scattering1D_filter_factory_T(backend)
-
